third_party/blink/web_tests/platform/generic/fast/url/path-url-expected.txt

Canonicalization of path URLs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS canonicalize('javascript:') is 'javascript:'
PASS canonicalize('JavaScript:Foo') is 'javascript:Foo'
PASS successfullyParsed is true

TEST COMPLETE
